Remove interface FirebaseDbService
